+++ src/components/input/NumericInput.js
@@ -11,6 +11,7 @@ class SpinnerControls extends PureComponent {
     const iconProps = {
       color: inverse ? 'teal' : 'neutral',
       element: 'button',
+      tabindex: '-1',
       tint: inverse ? 'lightest' : 'darkest',
       type: 'button',
     };